services/payment/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
pytest==8.3.2
requests==2.31.0
pytest-dependency==0.5.1
stripe==11.1.1
python-dotenv==1.0.1
pytest-mock==3.10.0